create history component
